20080609-1042

+++ emu/port/win-x11a.c
@@ -515,7 +515,7 @@ enum {
 
 typedef struct ICursor ICursor;
 struct ICursor {
-	ulong	inuse;
+	int	inuse;
 	int	modify;
 	int	hotx;
 	int	hoty;
